[Refactor] Bookmarks Fragment 리팩터링
1. 리사이클러뷰 어댑터 재활용을 바라보게 만듦.
2. 기타등등 짜잘한거 다 수정

+++ app/src/main/java/place/pic/ui/main/mypage/mywritings/MyWritingsFragment.kt
@@ -65,7 +65,7 @@ class MyWritingsFragment : Fragment() {
 
     private fun onMyWritingClick(placeGridItem: PlaceGridItem) {
         val intent = Intent(requireActivity(), DetailViewActivity::class.java)
-        intent.putExtra("placeIdx", placeGridItem.placeIdx)
+        intent.putExtra("placeIdx", placeGridItem.placeIdx.toLong())
         startActivityForResult(intent, 5000)
     }
